Update Last Location Cache to be valid for up to 12 hours from latest heartbeat with valid location (#822)
* Move LastLocation and LocationCache to own module

* Update LastLocation to be 12 hours from last heartbeat with validate location

* refactor last location tests
