+++ mobile_verifier/tests/integrations/main.rs
@@ -3,6 +3,7 @@ mod common;
 mod boosting_oracles;
 mod heartbeats;
 mod hex_boosting;
+mod last_location;
 mod modeled_coverage;
 mod rewarder_mappers;
 mod rewarder_oracles;
